% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average-lift.R
\name{average_lift}
\alias{average_lift}
\title{Calculate Average Lift}
\usage{
average_lift(x, ...)
}
\arguments{
\item{x}{A \code{LiftSimResults} object returned by \code{\link[=fit]{fit()}}.}

\item{...}{Passed to methods. See details.}
}
\value{
A single numeric value representing the mean lift.
}
\description{
An S7 generic for calculating the average lift from simulation results.
}
\details{
The following arguments are supported in methods:
\describe{
\item{na.rm}{
Should missing values be removed? Default is \code{FALSE}}
}
}
\examples{
\dontrun{
source <- S3DataSource()
model <- LiftAnalysis(data_source = source)
results <- fit(model)
average_lift(results)
}

}
\seealso{
\code{\link[=fit]{fit()}}, \link{LiftAnalysis}
}
