% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch Data from a Data Source}
\usage{
fetch_data(source, ...)
}
\arguments{
\item{source}{A data source object. Currently supports \link{S3DataSource}.}

\item{...}{Passed to methods}
}
\value{
A data.table containing the fetched data.
}
\description{
An S7 generic for fetching data from various sources. The default method
for \link{S3DataSource} downloads a CSV file from S3 and returns it as a
data.table.
}
\examples{
\dontrun{
Sys.setenv(S3_BUCKET_NAME = "my-bucket", S3_DATA_KEY = "data.csv")
source <- S3DataSource()
data <- fetch_data(source)
}

}
\seealso{
\link{S3DataSource}
}
