% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average-lift.R
\name{fit}
\alias{fit}
\title{Fit a Lift Model}
\usage{
fit(model, ...)
}
\arguments{
\item{model}{A model object. Currently supports \link{LiftAnalysis}.}

\item{...}{Passed to methods. See details.}
}
\value{
A \code{LiftSimResults} object (s3-classed numeric vector) containing the
simulated lift values.
}
\description{
An S7 generic for fitting the lift "model." The default method for
\link{LiftAnalysis} runs a "simulation" by scaling the spend column and adding
random noise.
}
\details{
The following arguments are supported in methods:
\describe{
\item{scaling_factor}{
Numeric scaling factor applied to spend values. Default is \code{1.5}.}
}
}
\examples{
\dontrun{
source <- S3DataSource()
model <- LiftAnalysis(data_source = source)
results <- fit(model, scaling_factor = 2.0)
}

}
\seealso{
\link{LiftAnalysis}, \code{\link[=average_lift]{average_lift()}}
}
